coin selection test + cleanup
