--- client/test-assets/DO_NOT_DELETE.md
@@ -1,1 +1,0 @@
-Do not delete files in this folder, those are used for testing.
