--- client/test-assets/wallet.data
@@ -1,1 +1,0 @@
-{"id":"dd49d56b-0223-482b-be6f-c8e412e5b2bc","keys_last_derived_pos":2,"master_priv_key":"tprv8ZgxMBicQKsPetp2zouqPnrzW7RG7ghGrTFrQnFa5msH86m99nax467wmprg8yv94tZtJpuUu8FhyMFWsD7Hu4U3TC1aFi6RKhSP1Z9RMM2","network":"regtest","se_backup_keys_last_derived_pos":2,"se_backup_keys_pos_encoded":"[]","se_key_shares_last_derivation_pos":1,"se_key_shares_pos_encoded":"[9999999]","se_proof_keys_last_derivation_pos":2,"se_proof_keys_pos_encoded":"[]","shared_keys":"[]"}
